# Speedup listing of all ions in database