--- a/fiasco/fiasco.py
+++ b/fiasco/fiasco.py
@@ -32,21 +32,23 @@ def list_elements(hdf5_dbase_root, sort=True):
     return elements
 
 
-@functools.lru_cache()
+#@functools.lru_cache()
 def list_ions(hdf5_dbase_root, sort=True):
     """
     List all available ions in the CHIANTI database
     """
-    ions = []
-    root = DataIndexer.create_indexer(hdf5_dbase_root, '/')
-    for f in root.fields:
-        try:
-            el = plasmapy.atomic.atomic_symbol(f.capitalize())
-            for i in root[f].fields:
-                if f == i.split('_')[0]:
-                    ions.append(f"{el} {i.split('_')[1]}")
-        except InvalidParticleError:
-            continue
+    root = DataIndexer(hdf5_dbase_root, '/')
+    ions = root['ion_index']
+    if ions is None:
+        ions = []
+        for f in root.fields:
+            try:
+                el = plasmapy.atomic.atomic_symbol(f.capitalize())
+                for i in root[f].fields:
+                    if f == i.split('_')[0]:
+                        ions.append(f"{el} {i.split('_')[1]}")
+            except InvalidParticleError:
+                continue
     # Optional because adds significant overhead
     if sort:
         ions = sorted(ions, key=lambda x: (plasmapy.atomic.atomic_number(x.split()[0]),