--- a/fiasco/util/setup_db.py
+++ b/fiasco/util/setup_db.py
@@ -113,6 +113,9 @@ def build_hdf5_dbase(ascii_dbase_root, hdf5_dbase_root, files=None):
                 progress.update()
             # Build an index for quick lookup of all ions in database
             from fiasco import list_ions  # import here to avoid circular imports
+            # Delete it if it already exists to ensure the index is rebuilt
+            if 'ion_index' in hf:
+                del hf['ion_index']
             ion_list = list_ions(hdf5_dbase_root)
             ds = hf.create_dataset('ion_index', data=np.array(ion_list).astype(np.string_))
             ds.attrs['unit'] = 'SKIP'