Addresses #69 #73 ping @dstansby 

This adds an `ion_index` entry at the top of the database. `fiasco.list_ions` will then just pull from this rather than searching the entire database tree to create the list.

Some comparisons when creating an `Ion`: 

* original (no index, no cache)
  ```python
  %timeit fiasco.base.IonBase('Fe 18')
  36.3 ms ± 327 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
  ```
* with #69 (i.e. LRU cache)
  ```python
  %timeit fiasco.base.IonBase('Fe 18',)
  113 µs ± 601 ns per loop (mean ± std. dev. of 7 runs, 10000 loops each)
  ```
* this PR
  ```python
  %timeit fiasco.base.IonBase('Fe 18',)
  1.25 ms ± 4.93 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
  ```

This is ~30x faster than the original method, but still about ~10x slower than the cache method. However, I can still create a collection of every ion in the database in <1 s using this PR so I'm not sure that further optimization is that necessary?