@dstansby 2020-03-08
The overhead in creating Ions really kicks in when every element and then subsequently every ion is created (~5000 I think, so 5000 calls to `Ion.__init___`). This PR seems like a sensible way to go 👍