@wtbarnes 2020-03-08
Out of curiosity, in what context are you creating 5000 ions? There are only about 495 unique ions in the database which is the example I mentioned in the above comment. If there are use cases for creating many more than this, the overhead may be more than my original estimate.