@dstansby 2020-03-08
Ah I think I'm an order of magnitude out - 500 sounds more like it... With your timings up top before this change it took 18 seconds to create all 500, which was a significant slow down to `level_populations`.