@wtbarnes 2020-03-08
Ok, I'll go ahead and merge this then. If you find it is still a bottleneck for some use cases, we could always go back to the lru_cache method.